更换护理列表接口，生成最新APK包

+++ app/src/main/java/com/yjkj/property_management/java/Entity/UserInfoEntity.kt
@@ -9,7 +9,7 @@ data class UserInfoEntity(
 data class UserInfoData(
     var createBy: Any?,
     var createTime: String?,
-    var markerId: Int,
+    var markerId: Long,
     var markerName: Any?,
     var ownerAccountName: Any?,
     var ownerAddress: String?,
